Jp5 2d457 1 des

+++ apply_patches.sh
@@ -3,15 +3,21 @@
 set -e
 
 if [[ $# < 1 ]]; then
-    echo "apply_patches.sh [--one-cam] apply [JetPack_version]"
+    echo "apply_patches.sh [--one-cam | --dual-cam] apply [JetPack_version]"
     echo "apply_patches.sh reset [JetPack_version]"
     exit 1
 fi
 
 # Default to dual camera DT for JetPack 5.0.2
-JP5_D4XX_DTSI="tegra194-camera-d4xx.dtsi"
+# single - dev board
+# one/dual - evb
+JP5_D4XX_DTSI="tegra194-camera-d4xx-dual.dtsi"
 if [[ "$1" == "--one-cam" ]]; then
     JP5_D4XX_DTSI="tegra194-camera-d4xx-single.dtsi"
+    shift
+fi
+if [[ "$1" == "--dual-cam" ]]; then
+    JP5_D4XX_DTSI="tegra194-camera-d4xx-dual.dtsi"
     shift
 fi
 
@@ -20,6 +26,11 @@ DEVDIR=$(cd `dirname $0` && pwd)
 . $DEVDIR/scripts/setup-common "$2"
 
 cd "$DEVDIR"
+
+# set JP4 devicetree
+if [[ "$JETPACK_VERSION" == "4.6.1" ]]; then
+    JP5_D4XX_DTSI="tegra194-camera-d4xx.dtsi"
+fi
 
 # NVIDIA SDK Manager's JetPack 4.6.1 source_sync.sh doesn't set the right folder name, it mismatches with the direct tar
 # package source code. Correct the folder name.
@@ -40,8 +51,8 @@ apply_external_patches $1 $KERNEL_DIR
 apply_external_patches $1 hardware/nvidia/platform/t19x/galen/kernel-dts
 
 if [ $1 = 'apply' ]; then
-    cp $DEVDIR/d4xx.c $DEVDIR/sources_$JETPACK_VERSION/kernel/nvidia/drivers/media/i2c/
-    cp $DEVDIR/$JP5_D4XX_DTSI $DEVDIR/sources_$JETPACK_VERSION/hardware/nvidia/platform/t19x/galen/kernel-dts/common/tegra194-camera-d4xx.dtsi
+    cp $DEVDIR/kernel/realsense/d4xx.c $DEVDIR/sources_$JETPACK_VERSION/kernel/nvidia/drivers/media/i2c/
+    cp $DEVDIR/hardware/realsense/$JP5_D4XX_DTSI $DEVDIR/sources_$JETPACK_VERSION/hardware/nvidia/platform/t19x/galen/kernel-dts/common/tegra194-camera-d4xx.dtsi
 elif [ $1 = 'reset' ]; then
-    rm $DEVDIR/sources_$JETPACK_VERSION/kernel/nvidia/drivers/media/i2c/d4xx.c
+    [[ -f $DEVDIR/sources_$JETPACK_VERSION/kernel/nvidia/drivers/media/i2c/d4xx.c ]] && rm $DEVDIR/sources_$JETPACK_VERSION/kernel/nvidia/drivers/media/i2c/d4xx.c
 fi
